(new) Depend directly on TypeScript, for now
I really want TypeScript to be per-project / optional. For now, that's untenable, due to the eslint parser *directly depending* on it (instead of failing gracefully.)

Maybe I'll open an Issue about it. Looks like all the eslint/typescript stuff is under flux right now *anyway*, with a pending major 1.0 release.